Refactor test packages to use 'modelsTest' for better organization + Resolved FlightTest fails
