+++ src/test/java/modelsTest/decoratorsTest/FlightDecoratorTest.java
@@ -1,4 +1,4 @@
-package models.decoratorsTest;
+package modelsTest.decoratorsTest;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
