+++ src/test/java/modelsTest/decoratorsTest/fuelDecoratorTest/FuelDecoratorTest.java
@@ -1,4 +1,4 @@
-package models.decoratorsTest.fuelDecoratorTest;
+package modelsTest.decoratorsTest.fuelDecoratorTest;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
