+++ src/test/java/modelsTest/decoratorsTest/radarDecoratorTest/RadarDecoratorTest.java
@@ -1,4 +1,4 @@
-package models.decoratorsTest.radarDecoratorTest;
+package modelsTest.decoratorsTest.radarDecoratorTest;
 
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertTrue;
