+++ src/test/java/modelsTest/flightTest/flightTypes/MilitaryFlightTest.java
@@ -1,4 +1,4 @@
-package models.flightTest.flightTypes;
+package modelsTest.flightTest.flightTypes;
 
 import models.flight.flighttypes.MilitaryFlight;
 import org.junit.jupiter.api.Test;
